+++ lib/flutter_jsonschema_builder.dart
@@ -6,3 +6,4 @@ export 'src/models/json_form_schema_style.dart';
 export 'src/models/schema_form_file.dart';
 export 'src/models/json_form_media.dart';
 export 'src/models/stepped_form_config.dart';
+export 'src/builder/stepped_form_widgets.dart';